refactor:  move Locale type
