+++ src/components/Language/Language.tsx
@@ -1,0 +1,1 @@
+export type Locale = 'en' | 'es' | 'fr' | 'ja' | 'zh' | 'ko'
